fix proto compiling
