+++ .github/workflows/conda.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Install dependencies
         run: conda install -y python conda-build "grpcio-tools<1.62" conda-verify
       - name: Compiling the prototypes
-        run: python python-libraries/nanover-core/setup.py compile_proto --proto-dir=protocol
+        run: python ./python-libraries/compile_proto.py --proto-dir=./protocol --python-dir=./python-libraries/nanover-core/src/nanover/protoco
       - name: Building the Conda packages
         run: |
           $build_command \
