+++ win_compile.ps1
@@ -38,7 +38,7 @@ announce "Installing python test requirements"
 python -m pip install -r ./python-libraries/requirements.test ${user_option}
 
 announce "Compiling proto files to python"
-python ./python-libraries/nanover-core/setup.py compile_proto
+python ./python-libraries/compile_proto.py --proto-dir=./protocol --python-dir=./python-libraries/nanover-core/src/nanover/protocol
 
 announce "Installing the python packages"
 python -m pip install ${edit_option} ${user_option} ./python-libraries/nanover-core/
